Fix copy script
